publisher/mp3tags/requirements.txt
eyeD3==0.9
pyyaml==5.3
click==7.0
